% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_scrapes.R
\name{wg_get_daily}
\alias{wg_get_daily}
\title{Scrapes daily data for a weather station}
\usage{
wg_get_daily(station, min_date, max_date, remDr)
}
\arguments{
\item{station}{A wunderground station e.g. 'EGNM'}

\item{min_date}{e.g. '2022-01-01'}

\item{max_date}{e.g. '2022-03-01'}

\item{remDr}{a Seleium remotedriver object}
}
\value{

}
\description{
Scrapes daily data for a weather station
}
\examples{
\dontrun{
wg_get_daily('EGNM', '2022-01-01', '2022-03-01', remDr)
}
}
