% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraper.R
\name{wg_get_data}
\alias{wg_get_data}
\title{Gets one month of daily data}
\usage{
wg_get_data(station = NULL, year = NULL, month = NULL, remDr)
}
\arguments{
\item{station}{a weather station code e.g. 'EGNM'}

\item{year}{numeric year}

\item{month}{numeric month}

\item{remDr}{a Selenium remotedriver object}
}
\value{

}
\description{
Gets one month of daily data
}
\examples{
\dontrun{
wg_get_data('EGNM', 2022, 01, remDr)
}
}
